Move InfluxDB store and ReadingsService into internal/measurement (#107)
* refactor: move InfluxDB store and ReadingsService into internal/measurement

Moves the InfluxDB Client implementation into internal/measurement/store_influx.go
and ReadingsService into internal/measurement/service.go. A DeviceFinder interface
decouples measurement from device, satisfied by a thin bridge in sensors.
sensors/influx.go and sensors/readings_service.go become thin wrappers.

Closes #99

* Extract internal/provisioning package from internal/sensors (#108)

* refactor: extract internal/provisioning package from internal/sensors

Moves provisioning store, postgres implementation, errors, and service
into internal/provisioning. Sensors package keeps thin aliases/wrappers
for backward compat.

Closes #100

* refactor: split provisioning.go into errors, store, service files


---------


* refactor: move ErrInfluxWrite to errors.go in measurement package


---------
